% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExif.R
\name{getExif}
\alias{getExif}
\title{Returns selected exifTags and pad any missing tags}
\usage{
getExif(files_to_read)
}
\arguments{
\item{files_to_read}{A vector of files to read}
}
\value{
tibble containing key exif data
}
\description{
Returns selected exifTags and pad any missing tags
}
