% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewExif.R
\name{getNewExif}
\alias{getNewExif}
\title{Append an existing exif data set with files that have been updated/added}
\usage{
getNewExif(files_to_read, existing_exif)
}
\arguments{
\item{existing_exif}{A tibble containing previously read exif data}

\item{path_to_read}{A vector of directories to search for files}
}
\value{

}
\description{
Append an existing exif data set with files that have been updated/added
}
