% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmplot.R
\name{ggmplot}
\alias{ggmplot}
\title{Title}
\usage{
ggmplot(
  coords_to_fit,
  maxzoom = 18,
  api_key = Sys.getenv("GOOGLE_MAPS_APIKEY"),
  xsize = 600,
  ysize = 600
)
}
\arguments{
\item{coords_to_fit}{either a vector or a tibble of coordinates named
GPSLatitude, GPSLongitude}

\item{maxzoom}{api range is from 1 (whole world) to 25 (building)}

\item{api_key}{api key for access to static maps}

\item{xsize}{image size to return}

\item{ysize}{image size to return}
}
\value{
a ggplot scaled to fit coords_to_fit with a static map background
}
\description{
Similar to gmplot but using ggmap package
}
