% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmestzoom.R
\name{gmestzoom}
\alias{gmestzoom}
\title{Estimates zoom parameter for google maps to fit points
Calculation source https://stackoverflow.com/questions/6048975/google-maps-v3-how-to-calculate-the-zoom-level-for-a-given-bounds}
\usage{
gmestzoom(ne_coord, sw_coord)
}
\arguments{
\item{ne_coord}{vector of longitude, latitude for north-east corner of box}

\item{sw_coord}{vector of longitude, latitude for south-west corner of box}

\item{minzm}{minimal value to return, deafult of 18}
}
\value{
integer value of zoom to use in a google maps api call
}
\description{
Estimates zoom parameter for google maps to fit points
Calculation source https://stackoverflow.com/questions/6048975/google-maps-v3-how-to-calculate-the-zoom-level-for-a-given-bounds
}
