% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modfiyExif.R
\name{modifyExif}
\alias{modifyExif}
\title{Applies a set of changes to fix exif data}
\usage{
modifyExif(original_exif, new_exif)
}
\arguments{
\item{original_exif}{A data frame of exif data}

\item{new_exif}{A data frame of exif data}
}
\value{
a data frame of status on changes
}
\description{
Applies a set of changes to fix exif data
}
