% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathtofiles.R
\name{pathtofiles}
\alias{pathtofiles}
\title{Helper function that converts a vector of paths to a vector of files in those directories}
\usage{
pathtofiles(paths_to_read)
}
\arguments{
\item{paths_to_read}{a vector of directories}
}
\value{
a vector of files in the given paths
}
\description{
Helper function that converts a vector of paths to a vector of files in those directories
}
