% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoTools-package.R
\docType{package}
\name{photoTools-package}
\alias{photoTools}
\alias{photoTools-package}
\title{photoTools: Tools to work with photos in R}
\description{
Functions and shiny apps to simplify manipulation and visualisation of exif data.
}
\author{
\strong{Maintainer}: Iain Bennett \email{iain.t.bennett@gmail.com}

}
\keyword{internal}
