% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shnyExifFixGPSLocal.R
\name{shnyExifFixGPSLocal}
\alias{shnyExifFixGPSLocal}
\title{Enables modification of exif data}
\usage{
shnyExifFixGPSLocal(
  exif_data,
  target_path,
  api_key = Sys.getenv("GOOGLE_MAPS_APIKEY"),
  gmap_data,
  lib_exif
)
}
\arguments{
\item{exif_data}{The data to modify}

\item{target_path}{The path to move processed files too}

\item{api_key}{An api key for google maps}

\item{gmap_data}{A data frame from prepGMH}

\item{lib_exif}{A data frame from getExif}
}
\value{
SHINY APP
}
\description{
Enables modification of exif data
}
\examples{
shnyViewer
}
