% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shnyExifFixTimeLocal.R
\name{shnyExifFixTimeLocal}
\alias{shnyExifFixTimeLocal}
\title{Enables modification of exif data}
\usage{
shnyExifFixTimeLocal(exif_data, target_path, existing_filenames)
}
\arguments{
\item{exif_data}{The data to modify}

\item{target_path}{The path to move processed files too}

\item{existing_filenames}{Any existing files that should be considered for UID}
}
\value{
SHINY APP
}
\description{
Enables modification of exif data
}
\examples{
shnyViewer
}
