% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shnyExifViewer.R
\name{shnyExifViewer}
\alias{shnyExifViewer}
\title{shnyViewer that shows EXIF data for an image}
\usage{
shnyExifViewer()
}
\value{
SHINY APP
}
\description{
shnyViewer that shows EXIF data for an image
}
\examples{
shnyViewer
}
