% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shnyExifViewerLocal.R
\name{shnyExifViewerLocal}
\alias{shnyExifViewerLocal}
\title{shnyViewer that shows EXIF data for an image}
\usage{
shnyExifViewerLocal(exif_data, apikey = Sys.getenv("GOOGLE_MAPS_APIKEY"))
}
\value{
SHINY APP
}
\description{
shnyViewer that shows EXIF data for an image
}
\examples{
shnyViewer
}
